use https proxy for lplib when set
This is a localised fix which passes Launchpad.login_anonymously's
proxy_info keyword argument a callable which can elicit the correct
Juju proxy configuration from the model for both http and https schemes.

This is inspired by the upstream httplib2 proxy_info_from_environment
method which is used in the launchpadlib implementation.